12-03-2023
Here is what I have finished so far. I also published a general overview ChatGPT helped me come up with on the Word doc. It all definitely needs more work. I will continue working on everything the morning.

Also, Justin, I noticed you made a terminal interface on one. (I think that was yours.) That would be awesome. It isn't in the instructions, but if it okay for us to do it that way, that would be much nicer.

What I am working on is just bare minimum to start.

-Morgan